requirements.txt
flask
discord.py
requests

gunicorn
